Made middleware changes for currentUser, and some required changes
The currentUser from the middleware is used to show the links on
navbar for logIn register or logout only when they are required
to be shown.
The other change done is that when we first register as a user,
we are still not able to access stuff that has the isLoggedIn
feature because this only works if we actually log in and not
just register. This has been fixed here so that we can access those
even when we register for the first time.