Beginning of shopping cart logic
Added a checkout link in navbar which is set to the length of
the array of the cart.
In show.ejs file we added another anchor tag to add products to
shopping cart.
Made a new route for cart. In this the logic for cart to add
products and increase the length of the cart array is done.